Move key_management schemas and modules into central datastore module
Part of centralizing migrations
